refactor: page 디렉토리 세부화
