+++ itshonigo/lib/pages/survey/confirm_survey.dart
@@ -1,7 +1,6 @@
-
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import '../main.dart';
+import '../../main.dart';
 
 class SurveyLast extends StatefulWidget {
